+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `ExpansionPanel`: fix children remaining in the DOM when closed
+
 ## [3.9.3][] - 2024-10-09
 
 ### Fixed